Fix: first attempt at fixing wakeupTriggered memory leak
When a SingleSubscriptionConsumerImpl is running for a long time, the listener
chain created on wakeupTriggered by the doPoll method using
ApiFuturesExtensions.whenFirstDone can grow to many GiB on the heap.

This is a quick-fix which makes sure that wakeupTriggered does not have
more Listeners attached on each poll request. But it also breaks any
early cancellation while waiting for the poll duration (when wakeup is
Triggered).